feat: Add script to generate random numbers
This commit adds two new Python scripts, `gen_ran_num.py` and `gen_ran_num2.py`, which generate a specified quantity of random numbers and write them to a file. The scripts use the current date and time to generate a unique filename for each run. The user can specify the quantity of random numbers to generate either by replacing the default value in the script or by entering it interactively when running `gen_ran_num2.py`.